Completes the Route guide Example with some adjustments
