--- routeguide/src/main/scala-task/ServerAppT.scala
@@ -1,40 +1,0 @@
-/*
- * Copyright 2017 47 Degrees, LLC. <http://www.47deg.com>
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package routeguide
-
-import freestyle.rpc.server.GrpcServerApp
-import freestyle.rpc.server.implicits._
-import journal.Logger
-import monix.eval.Task
-import monix.cats._
-import routeguide.serverT.implicits._
-
-import scala.concurrent.Await
-import scala.concurrent.duration.Duration
-
-object ServerAppT {
-
-  val logger: Logger = Logger[this.type]
-
-  def main(args: Array[String]): Unit = {
-
-    logger.info(s"Server is starting ...")
-
-    Await.result(server[GrpcServerApp.Op].bootstrapM[Task].runAsync, Duration.Inf)
-  }
-
-}
